refactor project structure and add containerization
