+++ server/.dockerignore
@@ -1,0 +1,6 @@
+.git
+.gitignore
+Dockerfile
+LICENSE
+Makefile
+README.md
