--- server/test.rest
@@ -1,12 +1,0 @@
-### POST REGISTER
-POST http://localhost:8000/register
-Content-Type: application/json
-
-{
-  "fullName": "Ulum",
-  "email": "ulumuddinihya022@gmail.com",
-  "password": "orakedisikanwongliyo22"
-}
-
-### GET ALL USER
-GET http://localhost:8000/users
